Refactor the Allocation class to a separate header
